onMount in header van button
- ONmount in het header component geplaatst
