+++ src/lib/molecules/header.svelte
@@ -1,8 +1,28 @@
 <script>
+  import { onMount } from 'svelte';
   import { page } from "$app/stores";
 
   let activeRoute;
   $: activeRoute = $page.route.id;
+
+  onMount(() => {
+        // console.log('js werkt')
+        let wrapper = document.querySelector('.wrapper')
+        let buttonThemeSwitcher = document.querySelector('.themeSwitcher')
+
+        buttonThemeSwitcher.addEventListener('click', themeSwitcher)
+
+        function themeSwitcher() {
+            console.log('werkt')
+            wrapper.classList.toggle('liquid-acid-theme')
+
+            if(this.innerHTML === "red pill") {
+                this.innerHTML = "blue pill"; // als er Frontend staat veranderd de innerHTML in Groen
+            } else {
+                this.innerHTML = "red pill"; // als er geen Frontend staat veranderd de innerHTML in Groen
+            }
+        }
+    });
 </script>
 
 <header>
