+++ src/routes/+layout.svelte
@@ -6,23 +6,6 @@
     onMount(() => {
         console.log('De component is gemonteerd');
         console.log('De hoogte van de box:', box.clientHeight);
-
-        // console.log('js werkt')
-        let wrapper = document.querySelector('.wrapper')
-        let buttonThemeSwitcher = document.querySelector('.themeSwitcher')
-
-        buttonThemeSwitcher.addEventListener('click', themeSwitcher)
-
-        function themeSwitcher() {
-            console.log('werkt')
-            wrapper.classList.toggle('liquid-acid-theme')
-
-            if(this.innerHTML === "red pill") {
-                this.innerHTML = "blue pill"; // als er Frontend staat veranderd de innerHTML in Groen
-            } else {
-                this.innerHTML = "red pill"; // als er geen Frontend staat veranderd de innerHTML in Groen
-            }
-        }
     });
 
     export let x = 0.0;
